refactor: reorganize getUrlFromContext
